Reduce running time of service/queue tests
The ordering relationships should be more-or-less invariant under changes to
the polling frequency, so I believe that the tests should be generally
testing the same thing even with much more frequent polls.

Frequent polling in production is wasteful of CPU, but for testing it means
that the tests complete more rapidly, so it seems a reasonable thing to do.